% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssrmst.R
\name{print.ssrmst}
\alias{print.ssrmst}
\title{print.ssrmst}
\usage{
\method{print}{ssrmst}(x, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments ignored in this function.}
}
\description{
S3 method for class 'ssrmst'
}

